Move test-constants to test directory
