+++ packages/server/src/__test__/integration/blockchain.test.ts
@@ -1,4 +1,4 @@
-import { DUMMY_RULES_ADDRESS } from '../../test-constants';
+import { DUMMY_RULES_ADDRESS } from '../../test/test-constants';
 import { Blockchain } from '../../wallet/services/blockchain';
 import { nitroAdjudicator } from '../../wallet/utilities/blockchain';
