+++ packages/server/src/test/test-constants.ts
@@ -1,6 +1,6 @@
 import { ethers } from 'ethers';
 import { Address, channelID, Uint256 } from 'fmg-core';
-import { HUB_ADDRESS } from './constants';
+import { HUB_ADDRESS } from '../constants';
 
 export const PARTICIPANT_1_PRIVATE_KEY =
   '0xa205281c09d630f6639c3505b63d57013996ba037bdbe4d2979eb8bd5bed5b1b';
